test: move stream and catalog cache tests to dedicated module
Keep test_stage_streaming.py unchanged from dev so parallel load-hardening
PRs merge without conflicts.
